+++ src/screens/Faucet/TokensFaucetTable.tsx
@@ -7,10 +7,9 @@ import { TableText } from "@components/Table";
 import Text, { TEXT_TYPES, TEXT_TYPES_MAP } from "@components/Text";
 import { FuelNetwork } from "@src/blockchain";
 import { SmartFlex } from "@src/components/SmartFlex";
+import { MINIMAL_ETH_REQUIRED } from "@src/constants";
 import { useMedia } from "@src/hooks/useMedia";
 import { useStores } from "@stores";
-
-import { MINIMAL_ETH_REQUIRED } from "../TradeScreen/RightBlock/CreateOrder/CreateOrder";
 
 import MintButtons from "./MintButtons";
