+++ src/screens/SpotScreen/BottomTables/BaseTable.tsx
