+++ src/screens/SpotScreen/BottomTables/BottomTables.tsx
