+++ src/screens/SpotScreen/BottomTables/SpotTable/SpotTable.tsx
