+++ src/screens/SpotScreen/BottomTables/SpotTable/SpotTableImpl.tsx
