+++ src/screens/SpotScreen/BottomTables/SpotTable/SpotTableVM.tsx
