+++ src/screens/SpotScreen/BottomTables/SpotTable/index.ts
