+++ src/screens/SpotScreen/BottomTables/constants.ts
