+++ src/screens/SpotScreen/Chart/ChartVm.tsx
