+++ src/screens/SpotScreen/Chart/TVChartContainer.tsx
