+++ src/screens/SpotScreen/Chart/TradingViewWidget.tsx
