+++ src/screens/SpotScreen/MarketStatistics.tsx
