+++ src/screens/SpotScreen/MarketStatisticsBar.tsx
