+++ src/screens/SpotScreen/OrderbookAndTradesInterface/SpotTrades/SpotTradesVM.tsx
