+++ src/screens/SpotScreen/RightBlock/CreateOrder/OrderTypeTooltip.tsx
