+++ src/screens/SpotScreen/RightBlock/MarketSelection/MarketTitle.tsx
