+++ src/screens/SpotScreen/RightBlock/MarketSelection/SpotMarketRow.tsx
