+++ src/screens/SpotScreen/RightBlock/RightBlock.tsx
