+++ src/screens/SpotScreen/RightBlock/index.ts
