+++ src/screens/SpotScreen/StatusBar/StatusBar.tsx
