+++ src/screens/SpotScreen/StatusBar/index.tsx
