+++ src/screens/SpotScreen/index.ts
@@ -1,0 +1,2 @@
+import SpotScreen from "./SpotScreen";
+export default SpotScreen;
