+++ src/screens/SwapScreen/SwapScreen.tsx
@@ -4,11 +4,10 @@ import styled from "@emotion/styled";
 import { observer } from "mobx-react";
 
 import { ModalEnums } from "@screens/SwapScreen/enums/modalEnums";
-import { MINIMAL_ETH_REQUIRED } from "@screens/TradeScreen/RightBlock/CreateOrder/CreateOrder";
 import ArrowDownIcon from "@src/assets/icons/arrowDown.svg?react";
 import { FuelNetwork } from "@src/blockchain";
 import Text, { TEXT_TYPES, TEXT_TYPES_MAP } from "@src/components/Text";
-import { DEFAULT_DECIMALS } from "@src/constants";
+import { DEFAULT_DECIMALS, MINIMAL_ETH_REQUIRED } from "@src/constants";
 import { useMedia } from "@src/hooks/useMedia";
 import { useWallet } from "@src/hooks/useWallet";
 import { useStores } from "@src/stores";
