--- src/screens/TradeScreen/index.ts
@@ -1,2 +1,0 @@
-import TradeScreen from "./TradeScreen";
-export default TradeScreen;
